auth/integration: refactor the bridge betwenn fpm tp fpm-auth
