+++ src/controller.rs
@@ -177,69 +177,3 @@ async fn fpm_ready(fpm_instance: &str, fpm_controller: &str) -> fpm::Result<()> 
     }
     Ok(())
 }
-
-// This API will be called from can_read and can_write functions
-//github-team:fpm-lang/ftd
-//github-starred:fpm-lang/ftd
-pub async fn get_remote_identities(
-    _remote_host: &str,
-    cookies: &std::collections::HashMap<String, String>,
-    identities: &[(String, String)],
-) -> fpm::Result<Vec<fpm::user_group::UserIdentity>> {
-    // use itertools::Itertools;
-
-    return dbg!(fpm::auth::github::get_auth_identities(cookies, identities).await);
-
-    // #[derive(serde::Deserialize)]
-    // struct UserIdentities {
-    //     success: bool,
-    //     reason: Option<String>,
-    //     #[serde(rename = "user-identities")]
-    //     user_identities: Option<Vec<std::collections::HashMap<String, String>>>,
-    // }
-    // let url = format!("https://{}/-/dj/get-identities/", remote_host);
-    // println!("remote url: {}", url);
-
-    /*
-    let cookie = cookies
-        .iter()
-        .map(|c| format!("{}={}", c.0, c.1))
-        .collect::<Vec<_>>()
-        .join(";");
-
-    println!("cookies: {}", cookie);
-
-    let headers = reqwest::header::HeaderMap::from_iter([(
-        reqwest::header::COOKIE,
-        reqwest::header::HeaderValue::from_bytes(cookie.as_bytes()).unwrap(),
-    )]);
-
-    let resp: UserIdentities =
-        crate::http::http_get_with_type(url::Url::parse(url.as_str())?, headers, identities)
-            .await?;
-
-    if !resp.success {
-        return Err(fpm::Error::APIResponseError(format!(
-            "url: {}, error while getting controller: get-identities",
-            url,
-        )));
-    }
-
-    #[allow(clippy::or_fun_call)]
-    let remote_identities = resp.user_identities.ok_or({
-        fpm::Error::APIResponseError(format!(
-            "controller:get-identities api error: {:?}",
-            resp.reason
-        ))
-    })?;
-
-    Ok(remote_identities
-        .into_iter()
-        .flat_map(|identities| {
-            identities
-                .into_iter()
-                .map(|(key, value)| fpm::user_group::UserIdentity { key, value })
-        })
-        .collect_vec())
-    */
-}
